third_party/WebKit/LayoutTests/editing/deleting/delete_after_span_ws.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
    assert_selection(
        '<div contenteditable><span>foo</span> |baz</div>',
        'delete',
        '<div contenteditable><span>foo|</span>baz</div>');

    assert_selection(
        '<div contenteditable><span>foo</span>    |baz</div>',
        'delete',
        '<div contenteditable><span>foo|</span>baz</div>');

    assert_selection(
        '<div contenteditable><span>foo</span>\n\n\n    |baz</div>',
        'delete',
        '<div contenteditable><span>foo|</span>baz</div>');
});
</script>
